backend/requirements.txt
bs4
pytz~=2024.1
python-dotenv~=1.0.1
fastapi~=0.111.0
pydantic
PyJWT==2.8.0
google-auth
google~=3.0.0
uvicorn~=0.30.3
aiohttp~=3.10.11
aiogram
motor
requests
pymongo
pymongo[srv]
langchain
langchain_core
langgraph
google-generativeai
langchain-google-genai
langsmith
langchain-community
langchain-google-community
starlette
eth-brownie

